Partial revert of previous commit
`getLedMode` accepts a `uint8_t&` parameter, which is a reference to a `uint8_t`. For this reason, you just pass in the variable, not a pointer to the variable. This is different to the `getTime` call because the signature of `getTime` changed from `getTime(uint8_t&, uint8_t&, uint8_t&)` to `getTime(uint8_t*, uint8_t*, uint8_t*)` when debugging, and was never changed back to the reference form.